Move to new Defaults directory, update docs
